+++ internal/cleanup/integration_test.go
@@ -37,15 +37,13 @@ func testClient(t *testing.T) *client.Client {
 
 	host := os.Getenv("TRUENAS_TEST_HOST")
 	apiKey := os.Getenv("TRUENAS_TEST_API_KEY")
-	socketPath := os.Getenv("TRUENAS_TEST_SOCKET")
 
 	// Live or Record mode
-	if host != "" || socketPath != "" {
+	if host != "" {
 		c, err := client.New(client.Config{
 			Host:               host,
 			APIKey:             apiKey,
 			InsecureSkipVerify: true,
-			SocketPath:         socketPath,
 		})
 		require.NoError(t, err)